bugfix: increase counter range
not just the temporary storage for it, but the actual couter too, needs
to be able to store more than 256 combinations... you're getting sloppy,
Kris!

+++ src/sleep.ino
@@ -118,7 +118,7 @@ void blink_signal(char msg) {
     burst_blink();
   }
 }
-byte counter = 0;
+int counter = 0;
 void loop()
 {
   // old way of sleeping:
